Map size/position related changes
Constrict map and picture on very narrow screens. Honor MINIMUM MAP
WIDTH, unless the photo get too small (less than MINIMUM MAP WIDTH), in
which case the difference is split.

Recenter the point if it is near the edge

Better calculations of center of visual map. Allows for map edge not
aligned with body